Enable macOS interactive mode
This is a WIP and not yet tested on macOS. The goal
here is to pair this with a modified JPype that starts
the JVM in a newly spawned Python thread. This would then
allow ImageJ's GUI to tie up the event loop without blocking
the main thread (in theory).